api/requirements.txt
annotated-types
bcrypt
bidict
blinker
chardet
click
dnspython
eventlet
Flask
Flask-Bcrypt
Flask-Cors
Flask-JWT-Extended
Flask-Mail
Flask-PyMongo
Flask-SocketIO
greenlet
h11
itsdangerous
Jinja2
MarkupSafe
motor
numpy
pandas
pillow
pydantic
pydantic_core
PyJWT
pymongo
python-dateutil
python-dotenv
python-engineio
python-slugify
python-socketio
pytz
reportlab
simple-websocket
six
text-unidecode
typing_extensions
tzdata
Werkzeug
wsproto
